Autoset future PAXes to be in future style
Automatically set PAXes with dates in the future to have the future marker
already set when you add it. The Future PAX style can still be changed as
needed, but now it will default.

The PAX being in the future is decided by hardcoded PAX dates compared
against what the browser reports as the current time.